==> code/07_longitudinal_analysis_v2.Rmd <==
---
title: "social nets  models"
output:
  html_document:
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: inline
---

This Rmd does all the mixed models using the data from recoding covariates longitudinal.Rmd

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(stargazer)
library(tidyverse)
library(kableExtra)
```


```{r make table}
#function to make table of coefficients
make_table <- function(model){
  coef <- summary(model)$coef[2]
  confint <- data.frame(confint(model))

  colnames(confint) = c("conf.low", "conf.high")

  table_out <- round(cbind(coef, confint[4,]),3)

  table_out <- table_out %>%
    unite(CI, conf.low, conf.high, sep = ",") %>%
    mutate(CI = paste0("(", CI, ")")) %>%
    unite(coef, coef, CI, sep = " ")
}

make_table_long <- function(model){
  coef_int_score <- summary(model)$coef[2]
  conf_age <- summary(model)$coef[3]
  conf_age2 <- summary(model)$coef[4]
  coef_interaction <- summary(model)$coef["int_score:years_over_65", "Estimate"]
  
  confint <- data.frame(confint(model))

  colnames(confint) = c("conf.low", "conf.high")
  
  coef_int_score_ci <- round(cbind("coef" = coef_int_score, confint[4,]),3)
  coef_int_score_age_ci <- round(cbind("coef" = conf_age, confint["poly(years_over_65, 2, raw = TRUE)1",]), 3)
  coef_int_score_age2_ci <- round(cbind("coef" = conf_age2, confint["poly(years_over_65, 2, raw = TRUE)2",]), 3)
  coef_interaction_ci <- round(cbind("coef" = coef_interaction, confint["int_score:years_over_65", ]), 3)
  
  table_out <- rbind(coef_int_score_ci, 
                     coef_int_score_age_ci, 
                     coef_int_score_age2_ci, 
                     coef_interaction_ci)

  table_out <- table_out %>%
    mutate(variable = c("Integration score", "Age", "Age2", "Integration score x age"), .before = "coef") %>%
    unite(CI, conf.low, conf.high, sep = ",") %>%
    mutate(CI = paste0("(", CI, ")")) %>%
    unite(coef, coef, CI, sep = " ")
}

make_table_long_conf <- function(model){
  coef_confi <- summary(model)$coef[2]
  conf_age <- summary(model)$coef["poly(years_over_65, 2, raw = TRUE)1", "Estimate"]
  conf_age2 <- summary(model)$coef["poly(years_over_65, 2, raw = TRUE)2", "Estimate"]
  coef_interaction <- summary(model)$coef["confidante_bin:years_over_65", "Estimate"]
  
  confint <- data.frame(confint(model))

  colnames(confint) = c("conf.low", "conf.high")
  
  coef_confi_ci <- round(cbind("coef" = coef_confi, confint[4,]),3)
  coef_age_ci <- round(cbind("coef" = conf_age, confint["poly(years_over_65, 2, raw = TRUE)1",]), 3)
  coef_age2_ci <- round(cbind("coef" = conf_age2, confint["poly(years_over_65, 2, raw = TRUE)2",]), 3)
  coef_interaction_ci <- round(cbind("coef" =coef_interaction, confint["confidante_bin:years_over_65", ]), 3)
  
  table_out <- rbind(coef_confi_ci, 
                     coef_age_ci, 
                     coef_age2_ci, 
                     coef_interaction_ci)

  table_out <- table_out %>%
    mutate(variable = c("Confidante", "Age", "Age2", "Confidante x age"), .before = "coef") %>%
    unite(CI, conf.low, conf.high, sep = ",") %>%
    mutate(CI = paste0("(", CI, ")")) %>%
    unite(coef, coef, CI, sep = " ")
}
```

## Datasets
```{r datasets}
# final_dat <- read.csv("/Users/ccalmasini/Desktop/Camilla KHANDLE/Social nets paper code/social_nets_dat.csv")
final_dat <- read.csv(here("social_nets_dat_both_waves.csv"))

#centering education at 12
#turning variables into factors
#dropping nas for complete case analysis
khandle1 <- final_dat %>%
  mutate(W1_EDU_new_c = W1_EDU_new - 12,
         W1_D_GENDER = factor(case_when(W1_D_GENDER == 1 ~ "Male",
                                        W1_D_GENDER == 2 ~ "Female")),
         W1_SMK = ifelse(W1_SMK == 0, "No", "Yes")) %>%
  drop_na(W1_EDU_new_c, income_num, ADL_IADL, tot_drinks_week, retirement_stat)

#turning data into long format
#each person is repeated 6 times: one for wave 1 execz, one for wave 1 vrmem, one for wave 1 semz, then same for wave 2
#make a r_d variable which combines race and cognitive domain
#then changing age so it coincides with age at either wave 1 or wave 2 depending
#lastly, age is centered at 65 and in decades
khandle1_long <- khandle1 %>%
  mutate(#income_num_c = scale(income_num, center = TRUE, scale = TRUE),
         income_num_c = log(income_num)) %>%
  gather(wave, cognitive_score, c(execz, execz_w2, vrmemz, vrmemz_w2, semz, semz_w2)) %>%
  mutate(cognitive_test = case_when(wave == "execz" | wave == "execz_w2" ~ "execz",
                                   wave == "semz" | wave == "semz_w2" ~ "semz",
                                   wave == "vrmemz" | wave == "vrmemz_w2" ~ "vrmemz"),
         wave = case_when(wave %in% c("execz", "semz", "vrmemz") ~ "wave1",
                          wave %in% c("execz_w2", "semz_w2", "vrmemz_w2") ~ "wave2")) %>%
  mutate(cognitive_test =  as.factor(cognitive_test),
         STUDYID = as.factor(STUDYID)) %>%
  unite(r_d, W1_D_RACE_SUMMARY, cognitive_test, remove = FALSE) %>%
  mutate(r_d = relevel(as.factor(r_d), ref = "Asian_execz")) %>%
  mutate(age = ifelse(wave == 'wave2', W2_INTERVIEW_AGE, W1_INTERVIEW_AGE),
         years_over_65 = (age - 65)/10)
```

## Integration score models using wave 1 and wave 2 data

```{r full sample main longitudinal}
#this is the longitudinal model
#i'm still including interaction between int_score and asian_vrmemz as before
#and also including the int_score x age interaction term
khandle1_long <- khandle1_long %>%
  mutate(asian_vrmemz = as.factor(ifelse(r_d == "Asian_vrmemz", 1, 0)))

#full sample model
m1_main_long <- lmer(cognitive_score ~ int_score + poly(years_over_65, 2, raw = TRUE) +
                       int_score:years_over_65 + r_d + int_score:asian_vrmemz + 
                       W1_D_GENDER + W1_EDU_new_c + (1|STUDYID), 
                     REML = FALSE, data = khandle1_long)

stargazer(summary(m1_main_long)$coef, 
          type = "text", 
          title = "Regression with interaction between linear age and int score")

#this just gives the coefficient for int_score with 95%CI
table1_long <- make_table_long(m1_main_long)
saveRDS(table1_long, 'output/tablem1_long.RDS')
```

### Stratified by race and using main covariates

```{r stratified, main covars}
#this does the longitudinal model by race
tables_main_covars <- list()
for(i in c("White", "Black", "LatinX", "Asian")){
  if (i != "Asian"){
    mod <- lmerTest::lmer(cognitive_score ~ int_score + poly(years_over_65, 2, raw = TRUE) + int_score:years_over_65 + 
                                        r_d + W1_D_GENDER + W1_EDU_new_c + (1 |STUDYID), 
                          REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  } else {
    mod <- lmerTest::lmer(cognitive_score ~ int_score + poly(years_over_65, 2, raw = TRUE) + int_score:years_over_65 + int_score:asian_vrmemz + 
                                        r_d + W1_D_GENDER + W1_EDU_new_c + (1 |STUDYID), 
                          REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  }
  
  stargazer(summary(mod)$coef, type = "text", title = paste("Regression for", i,  "participants, linear age interaction"))
  
  tables_main_covars[[i]] <- make_table_long(mod)
}

saveRDS(tables_main_covars, 'output/tables_main_covars_long.RDS')
```

### All covariates (ADL/IADL, alcohol, income and retirement status) - including wave 1 and wave 2 data

```{r full sample & all covs}
#this is full sample model including all covariates (ADL_IADL, alcohol, income, retirement status)
m2_long <- lmer(cognitive_score ~ int_score + poly(years_over_65, 2, raw = TRUE) +
                            int_score:years_over_65 + r_d + int_score:asian_vrmemz + W1_D_GENDER +
                            W1_EDU_new_c + ADL_IADL + tot_drinks_week + income_num_c +
                            retirement_stat + (1|STUDYID), 
                       REML = FALSE, data = khandle1_long)

stargazer(summary(m2_long)$coeff, 
          #m2_long, ci = T,
          type = "text", 
          title = "Linear interaction for full sample")

table2_long <- make_table_long(m2_long)
saveRDS(table2_long, 'output/tablem2_long.RDS')
```

### Stratified by race and using all covariates 

```{r stratified, all covariates}
#this does the longitudinal model by race but uses all covariates (ADL_IADL, alcohol, income, retirement status)
#i'm still including interaction between int_score and asian_vrmemz as before

tables_all_covars <- list()
for(i in c("White", "Black", "LatinX", "Asian")){
  if(i != "Asian"){
    mod_all <- lmerTest::lmer(cognitive_score ~ int_score + poly(years_over_65, 2, raw = TRUE) + int_score:years_over_65 + r_d + 
                                W1_D_GENDER + W1_EDU_new_c + ADL_IADL + tot_drinks_week + 
                                income_num_c + retirement_stat + (1|STUDYID), 
                       REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  
  } else {
    mod_all <- lmerTest::lmer(cognitive_score ~ int_score + poly(years_over_65, 2, raw = TRUE) + int_score:years_over_65 + 
                                int_score:asian_vrmemz + r_d + W1_D_GENDER + W1_EDU_new_c + ADL_IADL + tot_drinks_week + 
                                income_num_c + retirement_stat + (1|STUDYID), 
                       REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  }
  
  #class(mod_all) <- 'lmerMod'
  stargazer(summary(mod_all)$coef, 
            #mod_all, ci = T,
            type = "text", 
            title = paste("Regression for", i,  "participants, linear age interaction"))
  
  tables_all_covars[[i]] <- make_table_long(mod_all)
}

saveRDS(tables_all_covars, 'output/tables_all_covars_long.RDS')
```

## Confidante models including main covariates -- wave 1 and wave 2

This part of the code runs the models using confidante as the main independent variable instead of integration score

```{r main conf overall long}
#these are the models with confidante as exposure
#they should have an interaction between confidante and asian_vrmemz and confidante and latin_vrmemz
khandle1_long <- khandle1_long %>%
  mutate(latin_vrmemz = as.factor(ifelse(r_d == "LatinX_vrmemz", 1, 0)))

m1_confi <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + confidante_bin:asian_vrmemz+ 
                             confidante_bin:latin_vrmemz + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + W1_D_GENDER + W1_EDU_new_c + (1|STUDYID), 
                           REML = FALSE, data = khandle1_long)

stargazer(summary(m1_confi)$coeff, type = "text", title = "Confidante models with linear age interaction for full sample")

table1_confi <- make_table_long_conf(m1_confi)
saveRDS(table1_confi, 'output/table1_confi_long.RDS')
```

```{r stratified with main covariates}
#stratified
tables_main_confi <- list()
for (i in c("Asian", "Black", "LatinX", "White")){
  if(i == "Asian"){
    mod <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + 
                            confidante_bin:asian_vrmemz + 
                            W1_D_GENDER + W1_EDU_new_c + (1|STUDYID), 
                          REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  } else if (i == "LatinX"){
    mod <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + 
                            confidante_bin:latin_vrmemz + W1_D_GENDER + W1_EDU_new_c + (1|STUDYID), 
                          REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  } else {
    mod <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + 
                            W1_D_GENDER + W1_EDU_new_c + (1|STUDYID), 
                          REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  }
  
  summary(mod)$coef %>% round(3)
  confint(mod) %>% round(3)
  stargazer(summary(mod)$coef, type = "text", title = paste("Confidante models for", i, "with linear age interaction"))
  
  tables_main_confi[[i]] <- make_table_long_conf(mod)
}

saveRDS(tables_main_confi, 'output/tables_main_confi_long.RDS')
```

## Confidante models including enhanced covars - ADL_IADL, alcohol, income and retirement status 

```{r all covariates full sample}
#full sample
m2_confi <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + confidante_bin:asian_vrmemz + 
                                         confidante_bin:latin_vrmemz + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + 
                                          + W1_D_GENDER + W1_EDU_new_c + ADL_IADL + tot_drinks_week + income_num_c + retirement_stat + (1|STUDYID), 
                                       REML = FALSE, data = khandle1_long)

stargazer(summary(m2_confi)$coeff, 
          type = "text", 
          title = "Confidante models with linear age interaction for full sample -- all covars")

table2_confi <- make_table_long_conf(m2_confi)
saveRDS(table2_confi, 'output/table2_confi_long.RDS')
```

```{r stratified with all covariates}
#stratified
tables_all_confi <- list()
for (i in c("Asian", "Black", "LatinX", "White")){
  if(i == "Asian"){
    mod <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + confidante_bin:asian_vrmemz + 
                                          + W1_D_GENDER + W1_EDU_new_c + ADL_IADL + tot_drinks_week + income_num_c + retirement_stat + (1|STUDYID), 
                                       REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  } else if (i == "LatinX"){
    mod <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + confidante_bin:latin_vrmemz + 
                                         W1_D_GENDER + W1_EDU_new_c + ADL_IADL + tot_drinks_week + income_num_c + retirement_stat +(1|STUDYID), 
                                       REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  } else {
     mod <- lmerTest::lmer(cognitive_score ~ confidante_bin + r_d + poly(years_over_65, 2, raw = TRUE) + confidante_bin:years_over_65 + 
                                          W1_D_GENDER + W1_EDU_new_c + ADL_IADL + tot_drinks_week + income_num_c + retirement_stat +(1|STUDYID), 
                                       REML = FALSE, data = khandle1_long, subset = (W1_D_RACE_SUMMARY == i))
  }
  
  #summary(mod)$coef %>% round(3)
  #confint(mod) %>% round(3)
  #class(mod) <- 'lmerMod'
    stargazer(summary(mod)$coef, 
              #mod, ci = T, 
              type = "text", 
              title = paste("Confidante models for", i, "with linear age interaction"))

    #tables_all_confi[[i]] <- make_table_long_conf(mod)
}

saveRDS(tables_all_confi, 'output/tables_all_confi_long.RDS')
```
